test: split PDF equation component guard coverage
